# cardiogo-backend/app/docs/swagger.yml
swagger: "2.0"
info:
  title: CardioGo API
  description: >
    Documentación de la API de CardioGo.  
    **Autenticación con JWT:**  
    1. Ve al endpoint POST /auth/login.  
       Usa "Try it out" para enviar un email y contraseña de un usuario existente en tu base de datos.  
       Ejemplo:  
       email: juan.gonzalez@example.com  
       password: 123456  
       Copia el `access_token` recibido.  
    2. Haz clic en "Authorize" en Swagger y pega: `Bearer <access_token>`  
       ¡Listo! Ahora puedes probar endpoints protegidos.
  version: "1.0.0"
host: localhost:5000
basePath: /cardio-go/v1
schemes:
  - http
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: "Usa el token JWT con el prefijo `Bearer `, ejemplo: `Bearer eyJhbGciOiJIUzI1Ni...`"
paths:
...
  /users:
    get:
      tags:
        - Usuarios
      summary: Listar todos los usuarios
      security:
        - BearerAuth: []
      responses:
        200:
          description: Lista de usuarios
          schema:
            type: object
            properties:
              status:
                type: string
                example: success
              data:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    nombre:
                      type: string
                    apellido_paterno:
                      type: string
                    apellido_materno:
                      type: string
                    telefono:
                      type: string
                    email:
                      type: string
                    rol_id:
                      type: integer
    post:
      tags:
        - Usuarios
      summary: Crear nuevo usuario
      security:
        - BearerAuth: []
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              nombre:
                type: string
              apellido_paterno:
                type: string
              apellido_materno:
                type: string
              telefono:
                type: string
              email:
                type: string
              password:
                type: string
              rol_id:
                type: integer
      responses:
        201:
          description: Usuario creado exitosamente
          schema:
            type: object
            properties:
              status:
                type: string
                example: success
              data:
                type: object
                properties:
                  id:
                    type: integer
        400:
          description: Datos inválidos
...
  /alerts:
    get:
      tags:
        - Alertas
      summary: Listar todas las alertas
      security:
        - BearerAuth: []
      responses:
        200:
          description: Lista de alertas
          schema:
            type: object
            properties:
              status:
                type: string
                example: success
              data:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    tipo_alerta_id:
                      type: integer
                    nivel_alerta_id:
                      type: integer
                    paciente_id:
                      type: integer
                    fecha:
                      type: string
                      format: date-time
                    descripcion:
                      type: string
    post:
      tags:
        - Alertas
      summary: Crear nueva alerta
      security:
        - BearerAuth: []
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              tipo_alerta_id:
                type: integer
              nivel_alerta_id:
                type: integer
              paciente_id:
                type: integer
              fecha:
                type: string
                format: date-time
              descripcion:
                type: string
      responses:
        201:
          description: Alerta creada exitosamente
          schema:
            type: object
            properties:
              status:
                type: string
                example: success
              data:
                type: object
                properties:
                  id:
                    type: integer
  /alerts/{id}:
    get:
      tags:
        - Alertas
      summary: Obtener alerta por ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
      responses:
        200:
          description: Detalle de la alerta
          schema:
            type: object
            properties:
              status:
                type: string
                example: success
              data:
                type: object
                properties:
                  id:
                    type: integer
                  tipo_alerta_id:
                    type: integer
                  nivel_alerta_id:
                    type: integer
                  paciente_id:
                    type: integer
                  fecha:
                    type: string
                    format: date-time
                  descripcion:
                    type: string
        404:
          description: Alerta no encontrada

    put:
      tags:
        - Alertas
      summary: Actualizar alerta
      security:
        - BearerAuth: []
      consumes:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: integer
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              tipo_alerta_id:
                type: integer
              nivel_alerta_id:
                type: integer
              paciente_id:
                type: integer
              fecha:
                type: string
                format: date-time
              descripcion:
                type: string
      responses:
        200:
          description: Alerta actualizada exitosamente
          schema:
            type: object
            properties:
              status:
                type: string
                example: success
        404:
          description: Alerta no encontrada

    delete:
      tags:
        - Alertas
      summary: Eliminar alerta
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
      responses:
        200:
          description: Alerta eliminada correctamente
          schema:
            type: object
            properties:
              status:
                type: string
                example: success
        404:
          description: Alerta no encontrada
...
  /alert_levels:
    get:
      tags:
        - AlertLevels
      summary: List all alert levels
      security:
        - BearerAuth: []
      responses:
        200:
          description: List of alert levels
          schema:
            type: object
            properties:
              status:
                type: string
              data:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    nombre:
                      type: string
    post:
      tags:
        - AlertLevels
      summary: Create alert level
      security:
        - BearerAuth: []
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              nombre:
                type: string
      responses:
        201:
          description: Alert level created
          schema:
            type: object
            properties:
              status:
                type: string
              data:
                type: object
                properties:
                  id:
                    type: integer
  /alert_levels/{id}:
    get:
      tags:
        - AlertLevels
      summary: Get alert level by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
      responses:
        200:
          description: Alert level detail
          schema:
            type: object
            properties:
              status:
                type: string
              data:
                type: object
                properties:
                  id:
                    type: integer
                  nombre:
                    type: string
        404:
          description: Alert level not found
    put:
      tags:
        - AlertLevels
      summary: Update alert level
      security:
        - BearerAuth: []
      consumes:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: integer
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              nombre:
                type: string
      responses:
        200:
          description: Alert level updated
          schema:
            type: object
            properties:
              status:
                type: string
        404:
          description: Alert level not found
    delete:
      tags:
        - AlertLevels
      summary: Delete alert level
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
      responses:
        200:
          description: Alert level deleted
          schema:
            type: object
            properties:
              status:
                type: string
        404:
          description: Alert level not found
  /alert_types:
    get:
      tags:
        - AlertTypes
      summary: List all alert types
      security:
        - BearerAuth: []
      responses:
        200:
          description: List of alert types
          schema:
            type: object
            properties:
              status:
                type: string
              data:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    nombre:
                      type: string
    post:
      tags:
        - AlertTypes
      summary: Create alert type
      security:
        - BearerAuth: []
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              nombre:
                type: string
      responses:
        201:
          description: Alert type created
          schema:
            type: object
            properties:
              status:
                type: string
              data:
                type: object
                properties:
                  id:
                    type: integer
  /alert_types/{id}:
    get:
      tags:
        - AlertTypes
      summary: Get alert type by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
      responses:
        200:
          description: Alert type detail
          schema:
            type: object
            properties:
              status:
                type: string
              data:
                type: object
                properties:
                  id:
                    type: integer
                  nombre:
                    type: string
        404:
          description: Alert type not found
    put:
      tags:
        - AlertTypes
      summary: Update alert type
      security:
        - BearerAuth: []
      consumes:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: integer
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              nombre:
                type: string
      responses:
        200:
          description: Alert type updated
          schema:
            type: object
            properties:
              status:
                type: string
        404:
          description: Alert type not found
    delete:
      tags:
        - AlertTypes
      summary: Delete alert type
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
      responses:
        200:
          description: Alert type deleted
          schema:
            type: object
            properties:
              status:
                type: string
        404:
          description: Alert type not found